Pin all actions to hashes (#32)
For broadcast event, even though we write the code, it's not updated often
so it's not too much maintenance to pin it.

For actions/* from github we could choose not to pin them, because github
provides the runner and it's not possible to pin the runner. But it's
simpler to just pin everything, and we don't update them very often anyway,
and even without pinning you still have to update them for major releases.
If it prooves to be too much maintenance, we could unpin them or we could
use bots like dependabot or renovate to automate upgrading.

Signed-off-by: HARPER Jon <jon.harper87@gmail.com>