releases: run git clean before switching to main for the version bump commit
This removes build artifact from the branch (e.g. sonar reports) that could be
committed in the versionbump commit by mistake if the .gitignore is not
correctly setup and we add everything

Anecdotally, this allows prevents conflicts when switching which would fail the release
although the build should never modify tracked files.

Signed-off-by: HARPER Jon <jon.harper87@gmail.com>