backend-app, frontend-app: add base tag during releases for prettier git describe on main
Because we can release not at the tip of main but a few commits below, we never
have a tag on main corresponding to the point where we branched. As a simplest
solution, add a tag from now on, named a little bit like the release branch
(e.g. tag base-v2.18 like release-v2.18). An alternative would be to merge the
released tagged commit in main so that we have the real release tag in git
describe, but this means we don't have linear history. Not sure what is best.

Signed-off-by: HARPER Jon <jon.harper87@gmail.com>