backend-app, frontend-app: add next tag during releases for prettier git describe on main (#61)
Because we can release not at the tip of main but a few commits below, we never
have a tag on main corresponding to the point where we branched. As a simplest
solution, add a tag from now on, named a little bit like the release branch
(e.g. tag next-v2.18 like release-v2.18). An alternative would be to merge the
released tagged commit in main so that we have the real release tag in git
describe, but this means we don't have linear history. Not sure what is best.
We could add a tag "base-v2.18" at the commit where we branch for v2.18,
but then the git describe during the whole next iteration (preparing for v2.19)
will show v2.18. So instead, we add a tag "next-v2.19" at the versionbump commit

Signed-off-by: HARPER Jon <jon.harper87@gmail.com>